@@ -2,12 +2,12 @@
 
   <PropertyGroup>
     <OutputType>WinExe</OutputType>
-    <TargetFramework>net6.0-windows</TargetFramework>
+    <TargetFramework>net8.0-windows</TargetFramework>
     <UseWPF>true</UseWPF>
     <ApplicationIcon>FModel.ico</ApplicationIcon>
-    <Version>4.4.3.4</Version>
-    <AssemblyVersion>4.4.3.4</AssemblyVersion>
-    <FileVersion>4.4.3.4</FileVersion>
+    <Version>4.4.3.6</Version>
+    <AssemblyVersion>4.4.3.6</AssemblyVersion>
+    <FileVersion>4.4.3.6</FileVersion>
     <IsPackable>false</IsPackable>
     <IsPublishable>true</IsPublishable>
     <RuntimeIdentifier>win-x64</RuntimeIdentifier>
