@@ -1,30 +1,32 @@
-﻿using System;
+using System;
 using System.Linq;
-using System.Runtime.InteropServices;
+using System.Runtime.CompilerServices;
 using System.Windows;
 
 namespace FModel;
 
 public static class Helper
 {
-    [StructLayout(LayoutKind.Explicit)]
-    private struct NanUnion
-    {
-        [FieldOffset(0)]
-        internal double DoubleValue;
-        [FieldOffset(0)]
-        internal readonly ulong UlongValue;
-    }
-
     public static string FixKey(string key)
     {
         if (string.IsNullOrEmpty(key))
             return string.Empty;
 
-        if (key.StartsWith("0x"))
-            key = key[2..];
+        var keySpan = key.AsSpan().Trim();
+        if (keySpan.Length > sizeof(char) * (2 /* 0x */ + 32 /* FAES = 256 bit */)) // maybe strictly check for length?
+            return string.Empty; // bullshit key
 
-        return "0x" + key.ToUpper().Trim();
+        Span<char> resultSpan = stackalloc char[keySpan.Length + 2 /* pad for 0x */];
+        keySpan.ToUpperInvariant(resultSpan[2..]);
+
+        if (resultSpan[2..].StartsWith("0X"))
+            resultSpan = resultSpan[2..];
+        else
+            resultSpan[0] = '0';
+
+        resultSpan[1] = 'x';
+
+        return new string(resultSpan);
     }
 
     public static void OpenWindow<T>(string windowName, Action action) where T : Window
@@ -74,9 +76,9 @@ public static class Helper
 
     public static bool IsNaN(double value)
     {
-        var t = new NanUnion { DoubleValue = value };
-        var exp = t.UlongValue & 0xfff0000000000000;
-        var man = t.UlongValue & 0x000fffffffffffff;
+        var ulongValue = Unsafe.As<double, ulong>(ref value);
+        var exp = ulongValue & 0xfff0000000000000;
+        var man = ulongValue & 0x000fffffffffffff;
         return exp is 0x7ff0000000000000 or 0xfff0000000000000 && man != 0;
     }
 
@@ -96,13 +98,17 @@ public static class Helper
         return -d < n && d > n;
     }
 
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static float DegreesToRadians(float degrees)
     {
-        return MathF.PI / 180f * degrees;
+        const float ratio = MathF.PI / 180f;
+        return ratio * degrees;
     }
 
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static float RadiansToDegrees(float radians)
     {
-        return radians* 180f / MathF.PI;
+        const float ratio = 180f / MathF.PI;
+        return radians * ratio;
     }
 }
