@@ -13,9 +13,9 @@ public class CustomDirectory : ViewModel
             case "Fortnite [LIVE]":
                 return new List<CustomDirectory>
                 {
-                    new("Cosmetics", "FortniteGame/Content/Athena/Items/Cosmetics/"),
-                    new("Emotes [AUDIO]", "FortniteGame/Content/Athena/Sounds/Emotes/"),
-                    new("Music Packs [AUDIO]", "FortniteGame/Content/Athena/Sounds/MusicPacks/"),
+                    new("Cosmetics", "FortniteGame/Plugins/GameFeatures/BRCosmetics/Content/Athena/Items/Cosmetics/"),
+                    new("Emotes [AUDIO]", "FortniteGame/Plugins/GameFeatures/BRCosmetics/Content/Athena/Sounds/Emotes/"),
+                    new("Music Packs [AUDIO]", "FortniteGame/Plugins/GameFeatures/BRCosmetics/Content/Athena/Sounds/MusicPacks/"),
                     new("Weapons", "FortniteGame/Content/Athena/Items/Weapons/"),
                     new("Strings", "FortniteGame/Content/Localization/")
                 };
