@@ -16,14 +16,16 @@ public class GamePathVisualLineText : VisualLineText
     private ThreadWorkerViewModel _threadWorkerView => ApplicationService.ThreadWorkerView;
     private ApplicationViewModel _applicationView => ApplicationService.ApplicationView;
 
-    public delegate void GamePathOnClick(string gamePath);
+    public delegate void GamePathOnClick(string gamePath, string parentExportType);
 
     public event GamePathOnClick OnGamePathClicked;
     private readonly string _gamePath;
+    private readonly string _parentExportType;
 
-    public GamePathVisualLineText(string gamePath, VisualLine parentVisualLine, int length) : base(parentVisualLine, length)
+    public GamePathVisualLineText(string gamePath, string parentExportType, VisualLine parentVisualLine, int length) : base(parentVisualLine, length)
     {
         _gamePath = gamePath;
+        _parentExportType = parentExportType;
     }
 
     public override TextRun CreateTextRun(int startVisualColumn, ITextRunConstructionContext context)
@@ -56,14 +58,14 @@ public class GamePathVisualLineText : VisualLineText
         if (e.Handled || OnGamePathClicked == null)
             return;
 
-        OnGamePathClicked(_gamePath);
+        OnGamePathClicked(_gamePath, _parentExportType);
         e.Handled = true;
     }
 
     protected override VisualLineText CreateInstance(int length)
     {
-        var a = new GamePathVisualLineText(_gamePath, ParentVisualLine, length);
-        a.OnGamePathClicked += async gamePath =>
+        var a = new GamePathVisualLineText(_gamePath, _parentExportType, ParentVisualLine, length);
+        a.OnGamePathClicked += async (gamePath, parentExportType) =>
         {
             var obj = gamePath.SubstringAfterLast('.');
             var package = gamePath.SubstringBeforeLast('.');
@@ -80,17 +82,17 @@ public class GamePathVisualLineText : VisualLineText
                 }
                 else
                 {
-                    lineNumber = a.ParentVisualLine.Document.Text.GetLineNumber(obj);
+                    lineNumber = a.ParentVisualLine.Document.Text.GetNameLineNumber(obj);
                     line = a.ParentVisualLine.Document.GetLineByNumber(lineNumber);
                 }
-                
+
                 AvalonEditor.YesWeEditor.Select(line.Offset, line.Length);
                 AvalonEditor.YesWeEditor.ScrollToLine(lineNumber);
             }
             else
             {
                 await _threadWorkerView.Begin(cancellationToken =>
-                    _applicationView.CUE4Parse.ExtractAndScroll(cancellationToken, fullPath, obj));
+                    _applicationView.CUE4Parse.ExtractAndScroll(cancellationToken, fullPath, obj, parentExportType));
             }
         };
         return a;
